% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignepochs.R
\name{alignepochs}
\alias{alignepochs}
\title{Adjusts the values in a CF-compliant netCDF time coordinate
variable to match the given representation (baseline and units).}
\usage{
alignepochs(time, epoch = "days since 1950-01-01")
}
\arguments{
\item{time}{A CF-style time variable}

\item{epoch}{A CF time@units string}
}
\value{
A time variable with adjusted epoch and
}
\description{
The CF metadata standard requires a netCDF time coordinate
variable to be represented as time elapsed from an epoch, e.g.,
"days since 1972-01-01".  Different data sources may use different
epochs and units that need to be aligned with one another to
simplify analysis; this function will align them.
}
\details{
The \code{time} input variable must have an attribute named
"units" that follows the CF standard.  Currently supported time units for
conversion are "hours" and "days".

If the input variable also has a \code{calendar} attribute, that
will be respected in the adjustment; otherwise, it will assume the
data follows a standard (365.25-day) calendar.  Recognized
calendars are the same as those handled by the
\code{\link{yearlength}} function.
}
\examples{
time <- seq(365)-1
time@units <- "days since 1950-01-01"
time@calendar <- "noleap"
str(alignepochs(time, "hours since 1949-12-01"))

}
