% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.list.R
\name{as.matrix.list}
\alias{as.matrix.list}
\title{Converts a list to a matrix, padding vectors with NA to fit.}
\usage{
as.matrix.list(x, element = NULL, ...)
}
\arguments{
\item{x}{list to be converted to a matrix}

\item{element}{sub-element of each list to extract before
converting to matrix}

\item{...}{additional arguments (ignored; included for compatibility)}
}
\value{
a matrix
}
\description{
This function cbinds the elements of a list together into a
matrix.  Elements shorter than the longest element are padded with
NA values to fill out the matrix.  If the elments themselves are
lists, a specific sub-element can be extracted by name using the
\code{element} argument.
}
\details{
This is a convenience function intended primarily to make it easy
to overplot a list of similar objects all at once using
\code{matplot}.  The elements are padded out to a length of N by
indexing them from 1:N; this works sensibly for vectors, but may
produce unexpected results if the (sub-) elements of your list are
some other kind of object.
}
\examples{
a <- seq(5,10)
b <- rev(a)
N <- a * 100
x <- mapply(function(f,g){f + seq(g)/g}, a, N)
y <- mapply(rbeta, N, a, b)
z <- lapply(y, sort)

str(y)

matplot(as.matrix(x), as.matrix(z), pch=1)

d <- lapply(y, density, na.rm=TRUE)
matplot(as.matrix(d, "x"), as.matrix(d, "y"), type='l')

}
