% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsign.R
\name{atsign}
\alias{@}
\alias{@<-}
\alias{atsign}
\title{Overloaded infix @ for attribute access}
\usage{
object@name

object@name <- value
}
\arguments{
\item{object}{A non-S4 object.}

\item{name}{The character-string name of the attribute to access,
quoted or not.}

\item{value}{The value of the attribute to be set.}
}
\value{
The value of the attribute.
}
\description{
The at-sign, '@', is normally used with S4 objects for slot
access.  These functions allow the use of infix @ on non-S4
objects to get and set attributes.
}
\details{
The right-hand-side of the @ operator (the \code{name} argument)
need not be quoted.  It is converted to a literal character string
by computation on the language using \code{match.call}.  Attribute
names must be matched exactly.

These functions check whether their left-hand argument is an S4
object, and if so, pass it through to the \code{slot} function,
leaving the original use of @ for slot access unaffected.
}
\examples{

mu <- 3
sigma <- 2
x <- rnorm(10, mean=mu, sd=sigma)

attr(x, "mu") <- mu
print(attr(x, "mu"))

x@sigma <- sigma
print(x@sigma)

print(x)
x@mu <- NULL
x@mu_s <- mean(x)
x

}
\seealso{
\code{\link{copyatts}}
}

