% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{biascorrect}
\alias{biascorrect}
\title{Bias correction of climate model output using KDDM.}
\usage{
biascorrect(bcdata, norm = "zscore")
}
\arguments{
\item{bcdata}{The data to be bias-corrected.  A list with elements
named "obs", "cur", and "fut".}

\item{norm}{The type of normalization to use.  See
\code{\link{normalize}} for more details.}
}
\value{
The return value of the function.
}
\description{
This is the description of the function.  It's the first full
paragraph, and should briefly describe what the function does.
}
\details{
Third and subsequent paragraphs are details: a long section shown
after the argument that goes into detail about how the function
works.
}
\examples{
x <- "example code"
y <- "more example code"
paste(x, y)
}

