% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chauvenet.R
\name{chauvenet}
\alias{chauvenet}
\title{Trims extreme values from a vector that exceed Chauvenet's criterion.}
\usage{
chauvenet(x, upper = TRUE, lower = TRUE)
}
\arguments{
\item{x}{A vector of values to trim.}

\item{upper}{Logical; trim values in the upper tail.  Defaults to TRUE.}

\item{lower}{Logical; trim values in the lower tail.  Defaults to TRUE.}
}
\value{
A trimmed vector of values.
}
\description{
To prevent overfitting of the distributional tails when
constructing a \code{\link{distmap}} object, it is useful to trim
very extreme values from the input datasets.  This function trims
values whose probability is less than 1/N (greater than 1-1/N),
where N is the number of data points + 1/2, under the assumption that
the data is gaussian.
}
\details{
The probability bounds are calculated using \code{qnorm}.  Instead
of using the sample mean and standard deviation, this function
uses the median and the MAD (median absolute deviation), which are
more robust against the effects of very extreme values.
}
\examples{
x <- c(9, 10, 10, 10, 11, 50)
chauvenet(x)
chauvenet(c(x, -20), upper=FALSE)

set.seed(222)
y <- c(rnorm(100))
range(y)
range(chauvenet(y))
z <- c(y, 100, 1000)
qnorm(1-1/12, mean(z), sd(z))
range(chauvenet(z))
}
