% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climwindow.R
\name{climwindow}
\alias{climwindow}
\title{generate climatology moving window}
\usage{
climwindow(time, num = ceiling(year/inner), ratio = outer/inner,
  inner = year/num, outer = inner * ratio, year = yearlength(time))
}
\description{
creates moving window indexes for climatology iwth inner and outer
}

