% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof1d.R
\name{lof1d}
\alias{lof1d}
\title{Trims extreme values from a vector based on the local outlier
factor, simplified for the one-dimensional case.}
\usage{
lof1d(x, k = 5, cutoff = 3)
}
\arguments{
\item{x}{A vector of values.}

\item{k}{The number of nearest neighbors to consider.  Also the
number of upper tail values considered as possible outliers.}

\item{cutoff}{The maximum LOF for a value to be retained.}
}
\value{
The input vector \code{x} with any of the \code{k} largest
values that have an LOF greater than \code{cutoff} removed.
}
\description{
The local outlier factor (LOF) compares the local density of a
value to the average local density of its k nearest neighbors.  If
the ratio of the densities (the LOF) is significantly higher than
1, the point is an outlier and can be discarded.  This function
uses a highly simplified calculation to trim outlier values from
the upper tail of 1-dimensional data.
}
\details{
If length(x) is less than 3*k, the vector is returned unchanged.
See the Wikipedia page for a description of the LOF algorithm:
\url{https://en.wikipedia.org/wiki/Local_outlier_factor}
}
\examples{

set.seed(222)
x <- c(-5, rnorm(20), 2, 5, 10, 50, 100)
print(x)
print(lof1d(x))

}
