% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/namelist.R
\name{namelist}
\alias{namelist}
\title{Construct a list with automatic names}
\usage{
namelist(...)
}
\arguments{
\item{...}{objects to add to the list.}
}
\value{
A list with named elements
}
\description{
Constructs a lists from its arguments, automatically naming each
element of the list with the name of the argument.
}
\examples{
x <- 3
y <- "a string"
z <- function(x){x^3 +4}
n <- namelist(x,y,z)
str(namelist)
}

