% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nc_history.R
\name{nc_history}
\alias{nc_history}
\title{Adds a date-stamped comment to a NetCDF file's "history" attribute.}
\usage{
nc_history(nc, comment)
}
\arguments{
\item{nc}{An object of class 'ncdf4' (as returned by 'nc_open')
corresponding to the file whose history is to be updated.}

\item{comment}{The string to add to the history attribute.}
}
\description{
It's good practice to add a comment to a NetCDF file's "history"
attribute whenever the file's contents are changed.  This is a
convenience function that makes it easy to do so from R.
}
\details{
If a history attribute already exists, the comment is prepended to
it and separated by a carriage return.  A datestamp (produced by
\code{date()}) is added to the beginning of the comment.
}

