% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfskill.R
\name{pdfskill}
\alias{pdfskill}
\title{Calculates the overlap between two PDFs.}
\usage{
pdfskill(obs, mod, ...)
}
\arguments{
\item{obs}{A vector of observed data.}

\item{mod}{A vector of model data.}

\item{...}{Arguments to the density estimation function.}
}
\value{
A value between 0 and 1.
}
\description{
Given two datasets (observations and model), this function
calculates a skilll score as the area of overlap between their
estimated PDFs.  This metric comes from Perkins, et al. 2007 in
the Journal of Climate (doi: 10.1175/JCLI4253.1).
}
\details{
This implementation uses binned kernel density estimation from the
KernSmooth package to estimate the PDFs.
}
\examples{
obs <- rnorm(1000, mean=0, sd=1)
mod <- rnorm(1000, mean=0.3, sd=0.7)
pdfskill(obs, mod)

}
