% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.distmap.R
\name{quantile.distmap}
\alias{quantile.distmap}
\title{Calculates quantiles of one of the PDFs in a distmap object.}
\usage{
\method{quantile}{distmap}(x, probs = c(5, 10, 25, 50, 75, 90, 95)/100,
  which = c("xpdf", "ypdf"), ...)
}
\arguments{
\item{x}{An object of class 'distmap'.}

\item{probs}{A vector of probabilities.  Values outside [0,1] are
set to NA.  Defaults to percentiles useful for plotting distmap
objects.}

\item{which}{The name of the pdf (xpdf or ypdf) to use for
quantiles.  Defaults to "xpdf".}

\item{...}{Further arguments passed from other functions.}
}
\value{
The return value of the function.
}
\description{
This function does not use any of the quantile algorithms used by
the default quantile function.  Instead, it integrates the PDF in
the distmap object via the trapezoid rule to get a CDF, and then
inverts the CDF using splinefun.
}
\examples{
set.seed(222)
x <- rnorm(1000)
y <- rnorm(1000, mean=3, sd=2)
d <- distmap(x, y)
quantile(d)
quantile(d, which="ypdf", prob=c(25,50,75)/100)

}
