% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tailskill.R
\name{tailskill}
\alias{tailskill}
\title{Calculates the difference between PDF tails.}
\usage{
tailskill(obs, mod, threshold = 0.95, ...)
}
\arguments{
\item{obs}{A vector of observed data.}

\item{mod}{A vector of model data.}

\item{threshold}{The probability threshold where the tail begins.
Default value: 0.95.}

\item{...}{Arguments to the density estimation function.}
}
\value{
A numeric skill score; 1 indicates perfect agreement, 0 no
agreement.
}
\description{
Given two datasets (observations and model), this function
calculates a skill score for the agreement between the tails.  The
skill score is the area between the two PDFs above some threshold,
weighted linearly outwards.  This metric is based on Perkins, et
al. 2013 in the Int. J. Climatology (doi: 10.1002/joc.3500).
}
\details{
The upper tail is scored if the threshold value is >= 0.5;
otherwise, the lower tail is scored.  The \code{bkde} function
from the \code{KernSmooth} package is used to estimate the PDFs of
the two datasets.
}
\examples{
library(KernSmooth)
set.seed(22)
obs <- rgamma(1000, shape=1, scale=1)
perfect <- rgamma(1000, shape=1, scale=1)
good <- rgamma(1000, shape=6/5, scale=5/6)
bad <- rgamma(1000, shape=3, scale=1/3)

x <- namelist(obs, perfect, good, bad)

mplot(lapply(x, bkde), type="l", col=c("black","blue","green","red"), lty=1)

legend("topright", names(x), col=c("black","blue","green","red"), lty=1, lwd=2)

tailskill(obs, obs)
tailskill(obs, perfect)
tailskill(obs, good)
tailskill(obs, bad)
}

