% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzero.R
\name{unzero}
\alias{unzero}
\alias{unzero}
\alias{rezero}
\title{Remove precipitation zeros by setting them to NA}
\usage{
unzero(x, warn=FALSE)

rezero(x, refloor=TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{warn}{Logical; warn if NA values already present in input}

\item{refloor}{Logical; also set negative values to zero}
}
\description{
\code{unzero} eliminates zero precipitation data from the bias
correction by setting zero values to NA, which is ignored.
\code{rezero} sets NA values back to zero after bias correction.
}
\details{
The zeros in precipitation data cause problems for distribution
mapping.  Changing zeros to NA allows them to pass harmlessly
through the rest of the bias correction procedure without
requiring a lot of complicated and error-prone indexing to remove
and reinsert them from the arrays.

It is not uncommon for precipitation data to have small negative
values after bias correction.  By default \code{rezero} also sets
those values to zero when returning NA values back to zero.  If
the input has an attribute named "uncorrectable", \code{rezero}
will use it to index into the array and set those values to NA
afterwards.
}
\examples{

set.seed(22)

x <- rbinom(10, 2, 0.2)
y <- unzero(x)
x
y
rezero(y)
z <- unzero(c(x, NA, NA), warn=TRUE)
z <- c(z, -1)
z
rezero(z, refloor=FALSE)

}
