% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{biascorrect}
\alias{biascorrect}
\title{Bias correction of climate model output using KDDM.}
\usage{
biascorrect(bcdata, norm = "zscore", minobs = 250, dmap = FALSE, ...)
}
\arguments{
\item{bcdata}{The data to be bias-corrected.  A list with elements
named "obs", "cur", and "fut".}

\item{norm}{The type of normalization to use.  See
\code{\link{normalize}} for more details.}

\item{minobs}{The minimum number of data points to attempt a bias
correction.  If there are fewer that this many values in the
observations or in the current period model output (e.g., for
precipitation in a very dry location), the bias correction falls
back on a simple scaling by the ratio of obs mean to cur mean.
(If the ratio is undefined,  future values are passed through
unchanged.)}

\item{dmap}{Logical; if TRUE, returns the \code{\link{distmap}}
object generated during bias correction as an element named "dmap"
in the returned list.  N.B.: This option defaults to FALSE because
the distmap object is significantly larger than the inputs and
outputs of this function combined.}

\item{...}{Arguments to pass to \code{\link{distmap}}.}
}
\value{
The return value of the function.
}
\description{
This is the description of the function.  It's the first full
paragraph, and should briefly describe what the function does.
}
\details{
Third and subsequent paragraphs are details: a long section shown
after the argument that goes into detail about how the function
works.
}
\examples{
x <- "example code"
y <- "more example code"
paste(x, y)

}
