% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyatts.R
\name{copyatts}
\alias{copyatts}
\title{Copy all attributes from one variable to another}
\usage{
copyatts(from, to)
}
\arguments{
\item{from}{The variable to copy attributes from.}

\item{to}{The variable whose attribute are to be set.}
}
\value{
A copy of \code{to} with the attributes of \code{from}.
}
\description{
This function sets the attributes of the argument \code{to} to be
the same as the attributes of \code{from}.  Any existing
attributes are replaced.
}
\examples{
x <- 3
x@foo <- "things"
y <- 5
y@bar <- "more things"
y@baz <- "other things"
z <- copyatts(y,x)
z

}
