% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cslice.R
\name{cslice}
\alias{cslice}
\title{Climatological slicing of cyclic data}
\usage{
cslice(time, num = ceiling(year/inner), ratio = outer/inner,
  inner = year/num, outer = inner * ratio, year = yearlength(time),
  names = NULL)
}
\arguments{
\item{time}{A vector of times, encoded as time elapsed since some
start date.}

\item{num}{The number of windows.}

\item{ratio}{The length of the outer window as a multiple of the
length of the inner window.}

\item{inner}{The length of the inner window.}

\item{outer}{The length of the outer window.}

\item{year}{The length of the year.}

\item{names}{A vector of names for the windows.}
}
\value{
An object of class 'cslice' containing lists of indices
associated with moving windows across multiple years and a list of
the parameters specifying the object.
}
\description{
Creates an object that indexes climatological moving 
windows for cyclic data.
}
\details{
DETAILS GO HERE

Allows for irregularly spaced data and non-standard calendars.
Doesn't do anything with start time, so if they differ from file
to file, that could be weird.

The size and number of the inner and outer windows can be
specified using any two of the four parameters \code{num},
\code{ratio}, \code{inner}, and \code{outer}.  By default, the
\code{inner} and \code{outer} are assumed to have units of days,
but other units should work as long as they are consistent with
the value of \code{year}.
}
\seealso{
\code{\link{slice}}
}

