% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedrizzle.R
\name{dedrizzle}
\alias{dedrizzle}
\title{Removes excess drizzle from model data.}
\usage{
dedrizzle(x)
}
\arguments{
\item{x}{A list of vectors named "obs", "cur", and "fut".}
}
\value{
A list with obs, cur, and fut with drizzle removed.
}
\description{
Given three vectors of precipitation (current and future model
data, plus observations), dedrizzle finds a threshold value that
will equalize the wet/dry ratio in the model data to match the
wet/dry ratio in the observations, and sets all model values below
that threshold to zero.
}
\details{
Dedrizzle floors all its inputs at zero before calculation.  It
ignores NA values when calculating the wet/dry equalization
threshold.
}
\examples{
obs <- c(rep(0,10), rep(1,10), rep(NA,3))
cur <- c(seq(10)/10, NA)
fut <- seq(15)/10
namelist(obs,cur,fut)
dedrizzle(namelist(obs, cur, fut))

}

