% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Reversibly normalize a vector of values}
\usage{
normalize(x, norm = "zscore", mu = base::mean(x, na.rm = TRUE),
  sigma = stats::sd(x, na.rm = TRUE), lambda = 0, gamma = 0)
}
\arguments{
\item{x}{A vector of values to be transformed.}

\item{norm}{The type of normalization to apply.  See Details for
more information.  Type names can be abbreviated.}

\item{mu}{For zscore normalization, the mean.  Defaults to the
sample mean.}

\item{sigma}{For zscore normalization, the standard deviation.
Defaults to the sample standard deviation.}

\item{lambda}{For the Box-Cox transform normalization, the
exponent of the power transform. Defaults to zero, which is ideal
for gamma-distributed data.}

\item{gamma}{An offset to the data for the Box-Cox transform
normalization.  Must be greater than minus the minimum value,
otherwise NA values will result.  Defaults to zero.}
}
\value{
Both \code{normalize} and \code{denormalize} return a
vector of values.  \code{normalize} adds attributes to the vector
that record the type of normalization and the parameters used;
\code{denormalize} removes these attributes from its output.
}
\description{
Normalizes or standardizes a vector of values, recording the
details of the transformation in attributes of the result so that
the transformation can be reversed later on.
}
\details{
Three types of normalization are supported, chosen by the value of
the parameter \code{norm}:

\code{"zscore"}: subtract the mean (mu) and divide by the standard
deviation (sigma).  The result will have mean zero and unit
standard deviation.  Also known as normalizing residuals,
calculating the z-score, centering and scaling, or normalizing the
2nd central moment.  If the sample mean and sample variance are
used (as in the default arguments), this is technically
'studentizing' the data.

\code{"boxcox"}: apply the Box-Cox power transformation, which
raises the data to a power and scales it appropriately such that
it's continuous down to a power of 0, which becomes the log.  This
transformation will stabilize the variance for highly-skewed data.

\code{"identity"}: passes the data through unchanged.  This option
is sometimes useful for testing and development.

Each normalization by default uses parameters based on the sample
statistics of the input, but these parameters can be overridden.
}
\examples{

# z-score normalization
x <- rnorm(10000, mean=3, sd=2)
y <- normalize(x)
mplot(lapply(list(x,y),density), type="l")
str(y)

# Box-Cox normalization
x <- rgamma(10000, shape=3, rate=4)
y <- normalize(x, "boxcox")
mplot(lapply(list(x,y),density), type="l")
str(y)


}
\seealso{
\code{\link{denormalize}}
}

