% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Reversibly normalize a vector of values}
\usage{
normalize(x, norm = "zscore", mean = base::mean(x, na.rm = TRUE),
  sd = stats::sd(x, na.rm = TRUE), from = range(x, na.rm = TRUE),
  to = c(0, 1), scale = var(x, na.rm = TRUE)/mean(x, na.rm = TRUE),
  power = 0.25)
}
\arguments{
\item{x}{A vector of values to be transformed.}

\item{norm}{The type of normalization to apply.  See Details for
more information.  Type names can be abbreviated.}

\item{mean}{For zscore normalization, the mean.  Defaults to the
sample mean.}

\item{sd}{For zscore normalization, the standard deviation.
Defaults to the sample standard deviation.}

\item{from}{For range normalization, the input range.  Defaults to
range(x).}

\item{to}{For range normalization, the output range.  Defaults to
c(0,1).}

\item{scale}{For power normalization, the scaling factor to divide
by before the power transform.  Defaults to the sample variance
divided by the sample mean, which is the scale parameter of the
gamma and related distributions.}

\item{power}{For power normalization, the exponent of the power
transform. Defaults to 0.25, which is empirically known to be
well-suited for working with precipitation.}
}
\value{
Both \code{normalize} and \code{denormalize} return a
vector of values.  \code{normalize} adds attributes to the vector
that record the type of normalization and the parameters used;
\code{denormalize} removes these attributes from its output.
}
\description{
Normalizes or standardizes a vector of values, recording the
details of the transformation in attributes of the result so that
the transformation can be reversed later on.
}
\details{
Three types of normalization are supported, chosen by the value of
the parameter \code{norm}:

\code{"zscore"}: subtract the mean and divide by the standard
deviation.  The result will have mean zero and unit standard
deviation.  Also known as normalizing residuals, calculating the
z-score, centering and scaling, or normalizing the 2nd central
moment.  If the sample mean and sample variance are used (as in
the default arguments), this is technically 'studentizing' the
data.

\code{"range"}: subtract the minimum and divide by the range
(max-min).  The default is to transform the data to have min 0 and
max 1.

\code{"power"}: divide by the scale (variance/mean) and raise to a
power.  This transformation will stabilize the variance for
highly-skewed data.

Each normalization by default uses parameters based on the sample
statistics of the input, but these parameters can be overridden.
}
\examples{

# z-score normalization
x <- rnorm(10000, mean=3, sd=2)
y <- normalize(x)
mplot(lapply(list(x,y),density), type="l")
str(y)

# range normalization
x <- sample(2:9, 5, replace=TRUE)
y <- normalize(x, norm="range", from=c(1,10))
range(x)
range(y)
str(y)

# power normalization
x <- rgamma(10000, shape=3, rate=4)
y <- normalize(x, "power")
mplot(lapply(list(x,y),density), type="l")
str(y)


}
\seealso{
\code{\link{denormalize}}
}

