% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\alias{slice}
\alias{subslice}
\alias{unslice}
\title{Climatological slicing of timeseries}
\usage{
slice(x, how, outer=FALSE)

subslice(s, how, split=how$params$split)

unslice(s, how)
}
\arguments{
\item{x}{A vector of values to be sliced into windows.}

\item{how}{A \code{cslice} object defining the windows to use for
slicing.}

\item{outer}{Logical: if TRUE, uses the outer windows from
\code{how}, which can overlap.  If FALSE (the default), uses inner
windows, which cover the data completely without overlapping.}

\item{s}{Sliced data -- a list of vectors, each corresponding to
a different climatological window.}

\item{split}{Whether to segment subsliced data by year.  Defaults
to the \code{split} parameter of \code{how}.}
}
\description{
\code{slice}, \code{subslice}, and \code{unslice} are used to
separate a timeseries into climatological windows, then
reconstruct the timeseries from the windowed data.
}
\details{
\code{slice} uses a \code{\link{cslice}} object to separate a
timeseries into climatological windows.  \code{unslice}
reconstructs the timeseries from the (possibly modified) sliced
data.  \code{subslice} takes data that has been sliced into outer
windows and extracts the inner windows from each slice.

Whether the sliced data is further segmented by year depends on
whether the \code{cslice} object is segmented (i.e., called with
\code{split=TRUE}).  The \code{subslice} function also accepts a
\code{split} argument to override the default.
}
